separating json manager and game manager
